rename sideIconSidebar to tabbedSidebar
