+++ packages/ui/src/lib/tabbedSidebar/TabbedSidebar.svelte
