+++ packages/ui/src/lib/tabbedSidebar/TabbedSidebarTabLabel.svelte
