+++ packages/ui/src/lib/tabbedSidebar/TabbedSidebarTabList.svelte
